fix implicit search when using arbitrary
When using ArbitraryTypeReader, along with a custom ValueReader in a
class's companion object, the implicit in the companion needs to be
imported explicitly, or else the arbitrary implicit takes precedence.
This works around that by using a trick borrowed from circe of using a
wrapper type for the generated instances, with a conversion in
ValueReader's companion to the nested instance. This makes the arbitrary
instances lower priority than those in the companion.